refactor(companion): extract EventTypeDetailScreen to components/screens
Phase 2 of companion folder improvements:
- Extract EventTypeDetailScreen from app/event-type-detail.tsx to components/screens/
- Route file now serves as minimal wrapper with just Stack.Screen config
- Screen component receives route params as props
- Follows pattern established by BookingListScreen

This keeps /app folder focused on routing while moving screen logic to components.
